requirements.txt
torch
torchvision
scikit-learn
optuna
pyyaml
timm